refactor: "move configure-amplify to app dir"
