--- components/HomeHeader/styles.modules.css
@@ -1,9 +1,0 @@
-h1,
-h2,
-h3 {
-  line-height: 1.2em;
-}
-
-input {
-  height: 3em !important;
-}
